Refactor the code handling structures in Python
